renaming vblconfig to config.  doing some formatting changes.  Removed trailing_comma and trailing_semicolon because it requires nightly build chain.
