+++ rustfmt.toml
@@ -1,8 +1,7 @@
 # Rustfmt configuration for vestaboard-local
 # Note: Some features require nightly Rust toolchain
 
-# Enforce trailing commas in multi-line contexts (requires nightly)
-trailing_comma = "Vertical"
+
 
 # Basic formatting options
 max_width = 100
@@ -17,6 +16,5 @@ attr_fn_like_width = 70
 
 # Other stable formatting preferences
 match_block_trailing_comma = true
-trailing_semicolon = true
 use_field_init_shorthand = true
 use_try_shorthand = true
