+++ src/tests/mod.rs
@@ -3,6 +3,7 @@ mod api_broker_tests;
 mod api_tests;
 mod cli_display_tests;
 mod cli_setup_tests;
+mod config_tests;
 mod daemon_tests;
 mod error_tests;
 mod logging_tests;
